+++ models/utilities/constraint_check/src/constraint_test.cc
@@ -68,7 +68,7 @@ void
 ConstraintTest::count_violations()
 {
   if (violation != prev_violation) {
-    violation_count += violation;
+    violation_count += static_cast<unsigned int>(violation);
     prev_violation = violation;
   }
 }
